+++ scanpipe/pipelines/root_filesystems.py
@@ -42,9 +42,9 @@ class RootfsPipeline(Pipeline):
     @step
     def start(self):
         """
-        Load the Project instance.
+        Initialize the pipeline.
         """
-        self.project = self.get_project(self.project_name)
+        self.init_pipeline()
         self.next(self.find_root_filesystems)
 
     @step